% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind.R
\name{bind}
\alias{bind}
\title{Bind local variables to the return values of a function}
\usage{
bind(...)
}
\arguments{
\item{...}{List of variables to bind.}
}
\value{
An object containing the variable to be bound. This should be used
            in combination with the \code{\%<-\%} operator to bind values to
            parameters.
}
\description{
Bind return values from a function to local variables by combining the \code{bind}
function with the \code{\%<-\%} operator. When a function returns an object that is
a vector or list you will normally have to index into that result to get individual
components.
}
\details{
Using \code{bind} you can values in a vector or list to local variables, so

\code{bind(x, y, z) \%<-\% 1:3}

binds \code{a}, \code{b}, and \code{c} to values 1, 2, and 3, respectively.
}
\examples{
f <- function(x, y) c(x, y)
bind(a, b) \%<-\% f(1, 2)

g <- function(x, y) list(x, y)
bind(a, b) \%<-\% g(1, 2)

}

