% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bindr-package.R
\docType{package}
\name{bindr-package}
\alias{bindr}
\alias{bindr-package}
\title{Parametrized Active Bindings}
\description{
Provides a simple interface for creating active bindings where the
bound function accepts additional arguments.
}
\details{
See \code{\link[=create_env]{create_env()}} for creating an environment populated with active bindings,
and \code{\link[=populate_env]{populate_env()}} for populating an existing environment.
}

