% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind.R
\name{\%<-\%}
\alias{\%<-\%}
\title{Bind a `bind` object to values.}
\usage{
bindings \%<-\% value
}
\arguments{
\item{bindings}{A bindings object returned by the \code{\link{bind}} function.}

\item{value}{Values in a list/vector to be bound to the variables.}
}
\description{
Used to bind variables in the local scope to values, typically returned from
a function. A `bind` object is created using the \code{\link{bind}} function and
assigned to with this operator. Assigning this way will bind the parameters
specified in the \code{\link{bind}} call to the values on the right-hand-side
of the assignment.
}
\examples{
f <- function(x, y) c(x, y)
bind(a, b) \%<-\% f(1, 2)

g <- function(x, y) list(x, y)
bind(a, b) \%<-\% g(1, 2)

}
\seealso{
\code{\link{bind}}
}

